rename feature package
