+++ internal/cli/command/cluster/integratedservice/features/logging/activate.go
