+++ internal/cli/command/cluster/integratedservice/features/monitoring/common.go
