+++ internal/cli/command/cluster/integratedservice/features/monitoring/get.go
