+++ internal/cli/command/cluster/integratedservice/features/monitoring/spec.go
