+++ internal/cli/command/cluster/integratedservice/features/vault/common.go
