+++ internal/cli/command/cluster/integratedservice/features/vault/get.go
